created error directory in /api, cleaned up server.js
